+++ PinnyNotes.Core/Enums/NoteActions.cs
@@ -1,0 +1,10 @@
+﻿namespace PinnyNotes.Core.Enums;
+
+public enum NoteActions
+{
+    Created,
+    Updated,
+    Deleted,
+    Opened,
+    Closed
+}