+++ PinnyNotes.Core/Enums/NoteWindowActions.cs
@@ -1,0 +1,7 @@
+﻿namespace PinnyNotes.Core.Enums;
+
+public enum NoteWindowActions
+{
+    Open,
+    Close
+}