+++ PinnyNotes.WpfUi/Messages/MultipleNoteWindowActionMessage.cs
@@ -1,0 +1,8 @@
+﻿using PinnyNotes.Core.Enums;
+
+namespace PinnyNotes.WpfUi.Messages;
+
+public record MultipleNoteWindowActionMessage(
+    IEnumerable<int> NoteIds,
+    NoteWindowActions Action
+);