+++ PinnyNotes.WpfUi/Messages/OpenManagementWindowMessage.cs
@@ -1,0 +1,3 @@
+﻿namespace PinnyNotes.WpfUi.Messages;
+
+public record OpenManagementWindowMessage();