+++ PinnyNotes.WpfUi/Models/NotePreviewModel.cs
@@ -1,0 +1,52 @@
+﻿using System.Windows.Media;
+
+using PinnyNotes.Core.DataTransferObjects;
+using PinnyNotes.WpfUi.Themes;
+
+namespace PinnyNotes.WpfUi.Models;
+
+public class NotePreviewModel : BaseModel
+{
+    private const int MaxPreviewLength = 100;
+
+    public NotePreviewModel(NoteDto noteDto)
+    {
+        Id = noteDto.Id;
+
+        ContentPreview = noteDto.Content;
+
+        ThemeColorScheme = noteDto.ThemeColorScheme;
+    }
+
+    public int Id { get; set => SetProperty(ref field, value); }
+
+    public string ContentPreview
+    {
+        get;
+        set
+        {
+            string trimmedText = value.Trim();
+            int previewLength = Math.Min(trimmedText.Length, MaxPreviewLength);
+            string previewText = trimmedText[..previewLength];
+
+            SetProperty(ref field, previewText);
+        }
+    } = "";
+
+    public string ThemeColorScheme { get; set => SetProperty(ref field, value); }
+
+    public Brush BackgroundBrush { get; set => SetProperty(ref field, value); } = Brushes.LightGray;
+    public Brush BorderBrush { get; set => SetProperty(ref field, value); } = Brushes.DarkGray;
+    public Brush TitleBrush { get; set => SetProperty(ref field, value); } = Brushes.Gray;
+    public Brush TextBrush { get; set => SetProperty(ref field, value); } = Brushes.Black;
+
+    public bool IsSelected { get; set => SetProperty(ref field, value); }
+
+    public void UpdateBrushes(Palette palette)
+    {
+        BackgroundBrush = new SolidColorBrush(palette.Background);
+        BorderBrush = new SolidColorBrush(palette.Border);
+        TitleBrush = new SolidColorBrush(palette.Title);
+        TextBrush = new SolidColorBrush(palette.Text);
+    }
+}