+++ PinnyNotes.WpfUi/Services/WindowService.cs
@@ -16,19 +16,37 @@ public class WindowService
     private readonly AppMetadataService _appMetadataService;
     private readonly SettingsService _settingsService;
     private readonly NoteRepository _noteRepository;
-    private SettingsWindow? _settingsWindow;
+    private readonly ThemeService _themeService;
 
-    public WindowService(IServiceProvider serviceProvider, MessengerService messengerService, AppMetadataService appMetadataService, SettingsService settingsService, NoteRepository noteRepository)
+    private readonly Dictionary<int, NoteWindow> _openNoteWindows = [];
+    private SettingsWindow? _settingsWindow;
+    private ManagementWindow? _managementWindow;
+
+    public WindowService(
+        IServiceProvider serviceProvider,
+        MessengerService messengerService,
+        AppMetadataService appMetadataService,
+        SettingsService settingsService,
+        NoteRepository noteRepository,
+        ThemeService themeService
+    )
     {
         _serviceProvider = serviceProvider;
         _messengerService = messengerService;
         _appMetadataService = appMetadataService;
         _settingsService = settingsService;
         _noteRepository = noteRepository;
+        _themeService = themeService;
 
         _messengerService.Subscribe<ApplicationActionMessage>(OnApplicationActionMessage);
+
         _messengerService.Subscribe<OpenNoteWindowMessage>(OnOpenNoteWindowMessage);
+        _messengerService.Subscribe<MultipleNoteWindowActionMessage>(OnMultipleNoteWindowActionMessage);
+        _messengerService.Subscribe<NoteActionMessage>(OnNoteActionMessage);
+
         _messengerService.Subscribe<OpenSettingsWindowMessage>(OnOpenSettingsWindowMessage);
+
+        _messengerService.Subscribe<OpenManagementWindowMessage>(OnOpenManagementWindowMessage);
     }
 
     private void OnApplicationActionMessage(ApplicationActionMessage message)
@@ -37,7 +55,7 @@ public class WindowService
         switch (message.Action)
         {
             case ApplicationActions.Start:
-                _ = OpenNoteWindow();
+                OpenManagementWindow();
                 break;
             case ApplicationActions.NewInstance:
                 _ = OpenNoteWindow();
@@ -48,13 +66,30 @@ public class WindowService
     private void OnOpenNoteWindowMessage(OpenNoteWindowMessage message)
         => _ = OpenNoteWindow(message.NoteId, message.ParentNote);
 
-    private async Task OpenNoteWindow(int? noteId = null, NoteModel? parentNote = null)
+    private void OnMultipleNoteWindowActionMessage(MultipleNoteWindowActionMessage message)
     {
-        NoteViewModel viewModel = new(_noteRepository, _appMetadataService, _settingsService, _messengerService);
-        await viewModel.Initialize(noteId, parentNote);
-        NoteWindow newNote = new(_settingsService, _messengerService, viewModel);
+        foreach (int noteId in message.NoteIds)
+        {
+            switch (message.Action)
+            {
+                case NoteWindowActions.Open:
+                    _ = OpenNoteWindow(noteId);
+                    break;
+                case NoteWindowActions.Close:
+                    CloseNoteWindow(noteId);
+                    break;
+            }
+        }
+    }
 
-        newNote.Show();
+    private void OnNoteActionMessage(NoteActionMessage message)
+    {
+        switch (message.Action)
+        {
+            case NoteActions.Closed:
+                _openNoteWindows.Remove(message.NoteDto.Id);
+                break;
+        }
     }
 
     private void OnOpenSettingsWindowMessage(OpenSettingsWindowMessage message)
@@ -71,5 +106,52 @@ public class WindowService
             _settingsWindow.Show();
 
         _settingsWindow.Activate();
+    }
+
+    private void OnOpenManagementWindowMessage(OpenManagementWindowMessage message)
+        => OpenManagementWindow();
+
+    private async Task OpenNoteWindow(int? noteId = null, NoteModel? parentNote = null)
+    {
+        NoteWindow window;
+
+        if (noteId is not null && _openNoteWindows.TryGetValue((int)noteId, out NoteWindow? existingWindow))
+        {
+            window = existingWindow;
+            window.Activate();
+        }
+        else
+        {
+            NoteViewModel viewModel = new(_noteRepository, _appMetadataService, _settingsService, _messengerService, _themeService);
+            await viewModel.Initialize(noteId, parentNote);
+            window = new(_settingsService, _messengerService, _themeService, viewModel);
+
+            window.Show();
+
+            _openNoteWindows[viewModel.Note.Id] = window;
+        }
+    }
+
+    private void CloseNoteWindow(int noteId)
+    {
+        if (_openNoteWindows.TryGetValue(noteId, out NoteWindow? window) && window.IsLoaded)
+        {
+            window.Close();
+            _openNoteWindows.Remove(noteId);
+        }
+    }
+
+    private void OpenManagementWindow()
+    {
+        if (_managementWindow == null || !_managementWindow.IsLoaded)
+        {
+            _managementWindow = _serviceProvider.GetRequiredService<ManagementWindow>();
+            _managementWindow.Closed += (s, e) => _settingsWindow = null;
+        }
+
+        if (!_managementWindow.IsVisible)
+            _managementWindow.Show();
+
+        _managementWindow.Activate();
     }
 }